backend/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0
aiosqlite>=0.20.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
argon2-cffi>=23.1.0

# HTTP Client for API calls
httpx>=0.28.0
aiohttp>=3.11.0

# Validation & Settings (Python 3.14 compatible)
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.2.0

# Scheduling
apscheduler>=3.10.4

# Utilities
python-dotenv>=1.0.1
loguru>=0.7.2

# Webhooks
discord-webhook>=1.3.1
apprise>=1.9.0
